requirements.txt
torch
diffusers>=0.30.0
transformers
accelerate
safetensors
gradio>=4.0.0
Pillow
numpy
